requirement.txt
flask==2.1.1
flask-cores==3.0.10
scikit-image==0.19.2
scikit-learn==1.0.2
joblib==1.1.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1